requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.6
django-crispy-forms==1.7.2
idna==2.6
pkg-resources==0.0.0
pytz==2018.4
requests==2.18.4
urllib3==1.22
